refactor: ViewController에서 InitRef 삭제, InitRx => Init으로 변경
