+++ unity-skill-lab/Assets/Scripts/InGame/System/UIBindManager.cs
@@ -18,7 +18,7 @@ namespace InGame.System
         {
             foreach (var elem in _viewControllers)
             {
-                elem.InitRx(dataManager);
+                elem.Init(dataManager);
             }
         }
